+++ src/App.jsx
@@ -1,0 +1,12 @@
+import './App.css';
+import Input from './Input.jsx';
+
+function App() {
+  return (
+    <div className="App">
+      <Input />
+    </div>
+  );
+}
+
+export default App;
