+++ src/App.test.jsx
@@ -1,0 +1,8 @@
+import { render } from '@testing-library/react';
+import App from './App';
+
+test('renders learn react link', () => {
+  render(<App />);
+  
+  expect().toBeInTheDocument();
+});
