+++ src/Input.jsx
@@ -1,0 +1,31 @@
+import React from 'react';
+
+export default function Input({ value, onChange }) {
+  return (
+    <>
+    <p>
+      <label htmlFor="input-task-title">
+        LOG IN
+      </label>
+    </p>
+    <p>
+      <input
+        id="input-task-title"
+        type="text"
+        placeholder="ID"
+        value={value}
+        onChange={onChange}
+      />
+    </p>
+    <p>
+      <input
+        id="input-task-title"
+        type="text"
+        placeholder="password"
+        value={value}
+        onChange={onChange}
+      />
+      </p>
+      </>
+  );
+}
